# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  description: Logistics API provides services to find and manage entities for logistics
  title: Logistics API
  version: "1.0"
  contact:
    email: vladislavprokopenko1@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
...
  /cargos:
    get:
      tags:
        - Cargo
      summary: Find all Cargos
      operationId: cargosFindAll
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos were not found
        500:
          description: Internal Server Error

    post:
      tags:
        - Cargo
      summary: Create a new Cargo
      operationId: cargoCreate
      requestBody:
        description: Cargo object that needs to be added to cargo list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCargo'
      responses:
        201:
          description: Cargo was created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /cargos/byLogistician/{cargoId}:
    patch:
      tags:
        - Cargo
      summary: Update Cargo
      operationId: cargoUpdateByLogistician
      parameters:
        - name: cargoId
          in: path
          description: "Cargo id"
          required: true
          schema:
            type: integer
            format: int32
            example: 1001
      requestBody:
        description: Cargo object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCargoByLogistician'
      responses:
        200:
          description: Cargo was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /cargos/byDriver/{cargoId}:
    patch:
      tags:
        - Cargo
      summary: Update Cargo
      operationId: cargoUpdateByDriver
      parameters:
        - name: cargoId
          in: path
          description: "Cargo id"
          required: true
          schema:
            type: integer
            format: int32
            example: 1001
      requestBody:
        description: Cargo object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCargoByDriver'
      responses:
        200:
          description: Cargo was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /cargos/{cargoId}:
    delete:
      tags:
        - Cargo
      summary: Delete Cargo
      operationId: cargoDelete
      parameters:
        - name: cargoId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 1001
      responses:
        204:
          description: Cargo was deleted successfully
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This cargoId was not found
        500:
          description: Internal Server Error

    get:
      tags:
        - Cargo
      summary: Find Cargo by id
      operationId: cargoFindById
      parameters:
        - name: cargoId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 1001
      responses:
        200:
          description: Cargo was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This cargoId was not found
        500:
          description: Internal Server Error
  /cargos/byOrder/{orderForCargoId}:
    get:
      tags:
        - Cargo
      summary: Find Cargos by Order id
      operationId: cargoFindByOrderId
      parameters:
        - name: orderForCargoId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This cargoId was not found
        500:
          description: Internal Server Error
  /cargos/name/{name}:
    get:
      tags:
        - Cargo
      summary: Find Cargos by name
      operationId: cargoFindByCargoName
      parameters:
        - name: name
          in: path
          description: "Cargo name"
          required: true
          schema:
            type: string
            example: "Super Cola"
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with defined category were not found
        500:
          description: Internal Server Error
  /cargos/startCityAndState:
    get:
      tags:
        - Cargo
      summary: Find Cargos by start city and state
      operationId: cargoFindByStartCityAndState
      parameters:
        - in: query
          name: city
          description: Start city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Pittsburgh
          allowReserved: false
        - in: query
          name: state
          description: Start state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Pennsylvania
          allowReserved: false
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with defined start city and state were not found
        500:
          description: Internal Server Error
  /cargos/startCityAndStateAndAddress:
    get:
      tags:
        - Cargo
      summary: Find Cargos by start city, state and address
      operationId: cargoFindByStartCityAndStateAndAddress
      parameters:
        - in: query
          name: city
          description: Cargo start city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Pittsburgh
          allowReserved: false
        - in: query
          name: state
          description: Cargo start state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Pennsylvania
          allowReserved: false
        - in: query
          name: address
          description: Cargo start address
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: "2129 Bedford Ave"
          allowReserved: false
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with defined start address were not found
        500:
          description: Internal Server Error
  /cargos/nonLoaded:
    get:
      tags:
        - Cargo
      summary: "Find Cargos by status loaded = NO"
      operationId: cargoFindByNonLoaded
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Non-loaded cargos were not found
        500:
          description: Internal Server Error
  /cargos/loaded:
    get:
      tags:
        - Cargo
      summary: "Find Cargos by status loaded = YES"
      operationId: cargoFindByLoaded
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Loaded cargos were not found
        500:
          description: Internal Server Error
  /cargos/finalCityAndState:
    get:
      tags:
        - Cargo
      summary: Find Cargos by final city and state
      operationId: cargoFindByFinalCityAndState
      parameters:
        - in: query
          name: city
          description: Cargo final city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Columbus
          allowReserved: false
        - in: query
          name: state
          description: Cargo final state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Ohio
          allowReserved: false
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with defined final city and state were not found
        500:
          description: Internal Server Error
  /cargos/finalCityAndStateAndAddress:
    get:
      tags:
        - Cargo
      summary: Find Cargos by final city, state and address
      operationId: cargoFindByFinalCityAndStateAndAddress
      parameters:
        - in: query
          name: city
          description: Cargo final city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Columbus
          allowReserved: false
        - in: query
          name: state
          description: Cargo final state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Ohio
          allowReserved: false
        - in: query
          name: address
          description: Cargo final address
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: "480 E Broad St"
          allowReserved: false
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with defined final address were not found
        500:
          description: Internal Server Error
  /cargos/nonUnloaded:
    get:
      tags:
        - Cargo
      summary: "Find Cargos by status unloaded = NO"
      operationId: cargoFindByNonUnloaded
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Non-unloaded cargos were not found
        500:
          description: Internal Server Error
  /cargos/unloaded:
    get:
      tags:
        - Cargo
      summary: "Find Cargos by status unloaded = YES"
      operationId: cargoFindByUnloaded
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Unloaded cargos were not found
        500:
          description: Internal Server Error
  /cargos/earlier/expectedCompletionDateTime/{expectedCompletionDateTime}:
    get:
      tags:
        - Cargo
      summary: Find Cargos with expected completion date time earlier or equal to defined
      operationId: cargoFindByEarlierThanExpectedCompletionDateTime
      parameters:
        - name: expectedCompletionDateTime
          in: path
          description: "Expected completion date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with expected completion date time earlier or equal to defined were not found
        500:
          description: Internal Server Error
  /cargos/later/expectedCompletionDateTime/{expectedCompletionDateTime}:
    get:
      tags:
        - Cargo
      summary: Find Cargos with expected completion date time later or equal to defined
      operationId: cargoFindByLaterThanExpectedCompletionDateTime
      parameters:
        - name: expectedCompletionDateTime
          in: path
          description: "Expected completion date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with expected completion date time later or equal to defined were not found
        500:
          description: Internal Server Error
  /cargos/earlier/realCompletionDateTime/{realCompletionDateTime}:
    get:
      tags:
        - Cargo
      summary: Find Cargos with real completion date time earlier or equal to defined
      operationId: cargoFindByEarlierThanRealCompletionDateTime
      parameters:
        - name: realCompletionDateTime
          in: path
          description: "Real completion date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with real completion date time earlier or equal to defined were not found
        500:
          description: Internal Server Error
  /cargos/later/realCompletionDateTime/{realCompletionDateTime}:
    get:
      tags:
        - Cargo
      summary: Find Cargos with real completion date time later or equal to defined
      operationId: cargoFindByLaterThanRealCompletionDateTime
      parameters:
        - name: realCompletionDateTime
          in: path
          description: "Real completion date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Cargos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cargo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Cargos with real completion date time later or equal to defined were not found
        500:
          description: Internal Server Error
...
components:
  schemas:
...
    CreateCargo:
      description: Object representing Cargo creating
      type: object
      properties:
        orderForCargoId:
          type: integer
          format: int32
          example: 12001
        cargoName:
          type: string
          example: Super Cola
        weight:
          description: Cargo weight, tons
          type: number
          format: double
          example: 5.6
        startCity:
          type: string
          example: Pittsburgh
        startState:
          type: string
          example: Pennsylvania
        startAddress:
          type: string
          example: "2129 Bedford Ave"
        finalCity:
          type: string
          example: Columbus
        finalState:
          type: string
          example: Ohio
        finalAddress:
          type: string
          example: "480 E Broad St"
    UpdateCargoByLogistician:
      description: Object representing Cargo updating by Logistician
      type: object
      properties:
        cargoName:
          type: string
          example: Super Cola
        weight:
          description: Cargo weight, tons
          type: number
          format: double
          example: 5.6
        startCity:
          type: string
          example: Pittsburgh
        startState:
          type: string
          example: Pennsylvania
        startAddress:
          type: string
          example: "2129 Bedford Ave"
        finalCity:
          type: string
          example: Columbus
        finalState:
          type: string
          example: Ohio
        finalAddress:
          type: string
          example: "480 E Broad St"
    UpdateCargoByDriver:
      description: Object representing Cargo updating by Driver
      type: object
      properties:
        loaded:
          description: Loading status
          type: string
          enum: ["YES","NO"]
          default: "NO"
        unloaded:
          description: Unloading status
          type: string
          enum: ["YES","NO"]
          default: "NO"
    Cargo:
      description: Object representing Cargo
      type: object
      properties:
        cargoId:
          type: integer
          format: int32
          example: 1001
        orderForCargoId:
          type: integer
          format: int32
          example: 12001
        cargoName:
          type: string
          example: Super Cola
        weight:
          description: Cargo weight, tons
          type: number
          format: double
          example: 5.6
        startCity:
          type: string
          example: Pittsburgh,
        startState:
          type: string
          example: Pennsylvania
        startAddress:
          type: string
          example: "2129 Bedford Ave"
        loaded:
          description: Loading status
          type: string
          enum: ["YES","NO"]
          default: "NO"
        finalCity:
          type: string
          example: Columbus
        finalState:
          type: string
          example: Ohio
        finalAddress:
          type: string
          example: "480 E Broad St"
        unloaded:
          description: Unloading status
          type: string
          enum: ["YES","NO"]
          default: "NO"
        waypointIndex:
          description: Waypoint index based on Google Maps API optimization
          type: integer
          format: int32
          example: 1
        rideDistanceFromStartPoint:
          description: Ride distance from start point, km
          type: integer
          format: int32
          example: 355
        rideDurationFromStartPoint:
          description: Ride duration from start point, minutes
          type: integer
          format: int32
          example: 220
        rideDistanceFromPreviousPoint:
          description: Ride distance from previous point, km
          type: integer
          format: int32
          example: 216
        rideDurationFromPreviousPoint:
          description: Ride duration from previous point, minutes
          type: integer
          format: int32
          example: 123



        expectedCompletionDateTime:
          type: string
          format: date-time
        realCompletionDateTime:
          type: string
          format: date-time
...